# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
        "/api/create/applications": {
            "post": {
                "tags": [
                    "Application Section"
                ],
                "summary": "Request which create an application which automatically send an email to application creater",
                "description": "",
                "operationId": "9d62be4de8f19314028ca153f48497a5",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Please write application name",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Please write phone",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Please write password",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Please write text",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Please write email for send notifi.",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/applications": {
            "get": {
                "tags": [
                    "Application Section"
                ],
                "summary": "Request which returns all applications",
                "description": "",
                "operationId": "f681fcf248904a18edd68d2cc831051f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/delete/applications": {
            "post": {
                "tags": [
                    "Application Section"
                ],
                "summary": "Request which deletes certain application",
                "description": "",
                "operationId": "933a091fc452d345207e4c2b03b823f4",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "query",
                        "description": "Please write application_id",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}